// app/generated/fydao/accounts/donationRecord.ts
export type DonationRecord = {
  discriminator: ReadonlyUint8Array;
  bump: number;
  campaign: Address;
  donor: Address;
  /** Total amount this donor has deposited into the campaign escrow */
  amount: bigint;
};